added authorization

+++ app/Http/Controllers/Backend/CommunityController.php
@@ -61,6 +61,7 @@ class CommunityController extends Controller
      */
     public function edit(Community $community)
     {
+        $this->authorize('update', $community);
         return Inertia::render('Communities/Edit', [
             'community' => $community
         ]);
@@ -71,6 +72,7 @@ class CommunityController extends Controller
      */
     public function update(CommunityStoreRequest $request, Community $community)
     {
+        $this->authorize('update', $community);
         $community->update($request->validated());
 
         return to_route('communities.index')->with('message', 'Community updated successfully');
@@ -81,6 +83,7 @@ class CommunityController extends Controller
      */
     public function destroy(Community $community)
     {
+        $this->authorize('delete', $community);
         $community->delete();
         return back()->with('message', 'Community deleted successfully');
     }
